+++ api/src/subscribe/subscribe.module.ts
@@ -1,13 +1,12 @@
 import { Module } from '@nestjs/common';
-import { SubscribeController } from './subscribe.controller';
 import { SubscribeService } from './subscribe.service';
 import { SubscribeDAO } from './subscribe.dao';
 import { ContractModule } from '../contract/contract.module';
+import { SubscribeResolver } from './subscribe.resolver';
 
 @Module({
   imports: [ContractModule],
-  controllers: [SubscribeController],
-  providers: [SubscribeService, SubscribeDAO],
+  providers: [SubscribeService, SubscribeDAO, SubscribeResolver],
   exports: [SubscribeService],
 })
 export class SubscribeModule {}