+++ client/components/fetchAPI.js
@@ -1,14 +1,22 @@
-export const fetchAPI = async (input, init) => {
-    if (process.env.NEXT_PUBLIC_API_HOST){
-        return fetch(`${process.env.NEXT_PUBLIC_API_HOST}/api${input}`, {
-            ...(init ? init : {}),
-            headers: {
-                ...(init? init.headers : {}),
-            },
-            credentials: "include"
-        });
+export const fetchAPI = async (query, variables) => {
+    const params = {
+        method: "POST",
+        body: JSON.stringify({
+            query,
+            variables
+        }),
+        headers: {
+            Accept: "application/json",
+            "Content-Type": "application/json",
+        },
+        credentials: "include",
     }
-    return fetch(`/api${input}`, {
-        ...init,
-    });
+    let res = null;
+    if (process.env.NEXT_PUBLIC_API_HOST){
+        res = await fetch(`${process.env.NEXT_PUBLIC_API_HOST}/graphql`, params);
+    } else {
+        res = await fetch(`/graphql`, params);
+    }
+    const data = await res.json()
+    return data;
 }