+++ client/pages/dashboard.js
@@ -170,9 +170,17 @@ const Dashboard = (props) => {
             return
         }
         try {
-            const res = await fetchAPI("/user", { method: "GET" })
-            const data = await res.json()
-            setUsername(data.username)
+            const res = await fetchAPI(`
+            query getProfile{
+                getProfile{
+                    name
+                    role
+                }
+            }
+            `, {})
+            if (!res.errors) {
+                setUsername(res["data"]["getProfile"]["username"])
+            }
         } catch (err) {
             console.error(err)
         }
@@ -181,9 +189,22 @@ const Dashboard = (props) => {
     const getDashboardInfo = async () => {
         try {
             // fetch sub information by API
-            const res = await fetchAPI(`/subscribe/${page}`, { method: "GET" })
-            const data = await res.json()
-            let newSubs = data.subscriptions
+            const res = await fetchAPI(`
+            query getSubscriptions($input: Float!) {
+                getSubscriptions(page: $input) {
+                    subscriptions {
+                        emailAddrs
+                        contractAddrs
+                        pauseable
+                        createdAt
+                    }
+                    total
+                }
+            }
+            `, { input: page })
+            console.log(res)
+            const data = res.data["getSubscriptions"]
+            let newSubs = data["subscriptions"]
             const subAddrs = newSubs.map(({ contractAddr }) => contractAddr)
 
             // save total number of subscriptions